+++ getfile.h
@@ -1,3 +1,5 @@
+// last time modified by Igor UA3DJY on 20200124
+
 // -*- Mode: C++ -*-
 #ifndef GETFILE_H
 #define GETFILE_H
@@ -12,7 +14,7 @@ int ptt(int nport, int ntx, int* iptt, int* nopen);
 
 extern "C" {
 int ptt_(int nport, int ntx, int* iptt, int* nopen);
-void wav12_(short d2[], short d1[], int* nbytes, short* nbitsam2);
+void wav12_(int d2[], int d1[], int* nbytes, int* nbitsam2);
 }
